+++ src/ScottPlot4/changelog.md
@@ -3,6 +3,8 @@
 ## ScottPlot 4.1.57
 _not yet published on NuGet..._
 * Scatter: Improved `GetPointNearest()` when `OnNaN` is `Gap` or `Ignore` (#2048) _Thanks @thopri_
+* Heatmap and Image: Added `Coordinate[] ClippingPoints` to give users the ability to clip to an arbitrary polygon (#2049, #2052) _Thanks @xichaoqiang_
+* Image: Improved automatic axis limits measurement when `HeightInAxisUnits` is defined
 
 ## ScottPlot 4.1.56
 _Published on [NuGet](https://www.nuget.org/profiles/ScottPlot) on 2022-08-16_